More tests updated
