--- test/dom/helpers.js
@@ -1,6 +1,0 @@
-
-
-/**
- * Component Helpers
- */
-
