+++ test/dom/update-attributes.js
@@ -1,146 +1,86 @@
+import trigger from 'trigger-event'
+import raf from 'component-raf'
+import assert from 'assert'
+import {component,dom,scene} from '../../'
+import {mount} from '../helpers'
 
-var trigger = require('trigger-event');
-var raf = require('component-raf');
+var AttrComponent = component(function(props, state){
+  var attrs = {};
+  if (props.name) attrs.name = props.name;
+  return dom('span', attrs);
+});
 
-describe('Patching Attributes', function(){
-  var scene;
+it('should add/update/remove attributes', function(){
+  var app = scene(AttrComponent)
+  mount(app, function(el, renderer){
+    assert.equal(el.innerHTML, '<span></span>')
+    app.setProps({ name: 'Bob' })
+    renderer.render()
+    assert.equal(el.innerHTML, '<span name="Bob"></span>')
+    app.setProps({ name: 'Tom' })
+    renderer.render()
+    assert.equal(el.innerHTML, '<span name="Tom"></span>')
+    app.setProps({ name: null })
+    renderer.render()
+    assert.equal(el.innerHTML, '<span></span>')
+  })
+})
 
-  var AttrComponent = component(function(props, state){
-    var attrs = {};
-    if (props.name) attrs.name = props.name;
-    return dom('span', attrs);
+it('should not touch the DOM if attributes have not changed', function(){
+  var pass = true;
+  var app = scene(AttrComponent)
+    .setProps({ name: 'Bob' })
+  mount(app, function(el, renderer){
+    el.setAttribute = function(){
+      pass = false;
+    }
+    app.setProps({ name: 'Bob' })
+    renderer.render()
+    assert(pass)
+  })
+})
+
+it('should update the value of input fields', function () {
+  var Input = component({
+    render: function(props, state){
+      return dom('input', { value: props.value })
+    }
   });
 
-  beforeEach(function () {
-    scene = AttrComponent.render(window.el);
-    scene.update();
-  });
+  var app = scene(Input)
+    .setProps({ value: 'Bob' })
 
-  afterEach(function () {
-    scene.remove();
-  });
-
-  it('should have default state', function () {
-    assert.equal(el.innerHTML, '<span></span>');
-  });
-
-  it('should add attributes', function(){
-    scene.setProps({ name: 'Bob' });
-    scene.update();
-    assert.equal(el.innerHTML, '<span name="Bob"></span>');
-  })
-
-  it('should update attributes', function(){
-    scene.setProps({ name: 'Bob' });
-    scene.update();
-    scene.setProps({ name: 'Tom' });
-    scene.update();
-    assert.equal(el.innerHTML, '<span name="Tom"></span>');
-  })
-
-  it('should remove attributes', function(){
-    scene.setProps({ name: 'Bob' });
-    scene.update();
-    scene.setProps({ name: null });
-    scene.update();
-    assert.equal(el.innerHTML, '<span></span>');
-  })
-
-  it('should not update attributes that have not changed', function(){
-    var pass = true;
-    var Page = component({
-      afterMount: function(el){
-        el.setAttribute = function(){
-          pass = false;
-        }
-      },
-      render: function(props, state){
-        return dom('div', { name: props.name })
-      }
-    });
-    var page = Page.render(el, { name: 'Bob' });
-    page.setProps({ name: 'Bob' });
-    page.update();
-    assert(pass);
-    page.remove();
-  })
-
-  it('should update the value of input fields', function () {
-    var Input = component({
-      render: function(props, state){
-        return dom('input', { value: props.value })
-      }
-    });
-    var page = Input.render(el, { value: 'Bob' });
-    page.update();
+  mount(app, function(el, renderer){
     assert(el.querySelector('input').value === 'Bob');
     assert(el.querySelector('input').getAttribute('value') === 'Bob');
-    page.setProps({ value: 'Tom' });
-    page.update();
+    app.setProps({ value: 'Tom' });
+    renderer.render();
     assert(el.querySelector('input').value === 'Tom');
     assert(el.querySelector('input').getAttribute('value') === 'Bob');
-    page.remove();
   })
+})
 
-  it('should update element if function context changes', function(done){
-    var values = [];
-    var Page = component(function(props){
-      function push() {
-        values.push(props.value);
-      }
+it('should render innerHTML', function () {
+  var Test = component(function(){
+    return dom('div', { innerHTML: 'Hello <strong>World</strong>' });
+  });
+  var app = scene(Test)
+  mount(app, function(el, renderer){
+    assert.equal(el.innerHTML,'<div>Hello <strong>World</strong></div>')
+  })
+})
 
-      return Button({
-        label: 'A static prop',
-        onClick: push
-      });
-    });
-
-    var Button = component(function(props){
-      return dom('div.Button', { onClick: props.onClick }, [ props.label ]);
-    });
-
-    var page = Page.render(el, { value: 'One' });
-    page.update();
-    trigger(el.querySelector('.Button'), 'click');
-    page.setProps({ value: 'Two' });
-    page.update();
-    raf(function(){
-      trigger(el.querySelector('.Button'), 'click');
-      assert.equal(2, values.length);
-      assert.equal('One', values[0]);
-      assert.equal('Two', values[1]);
-      done();
-    });
+it('should update innerHTML', function () {
+  var Test = component(function(props){
+    return dom('div', { innerHTML: props.content });
   });
 
-  describe('innerHTML', function () {
+  var app = scene(Test)
+    .setProps({ content: 'Hello <strong>World</strong>' })
 
-    beforeEach(function(){
-      el.innerHTML = "";
-    })
-
-    it('should render innerHTML', function () {
-      var Test = component(function(){
-        return dom('div', { innerHTML: 'Hello <strong>World</strong>' });
-      });
-      var page = Test.render(el);
-      page.update();
-      assert.equal(el.innerHTML,'<div>Hello <strong>World</strong></div>');
-      page.remove();
-    })
-
-    it('should update innerHTML', function () {
-      var Test = component(function(props){
-        return dom('div', { innerHTML: props.content });
-      });
-      var page = Test.render(el, { content: 'Hello <strong>World</strong>' });
-      page.update();
-      page.setProps({ content: 'Hello <strong>Pluto</strong>' });
-      page.update();
-      assert.equal(el.innerHTML,'<div>Hello <strong>Pluto</strong></div>');
-      page.remove();
-    })
-
-  });
-
-});
+  mount(app, function(el, renderer){
+    app.setProps({ content: 'Hello <strong>Pluto</strong>' });
+    renderer.render();
+    assert.equal(el.innerHTML,'<div>Hello <strong>Pluto</strong></div>');
+  })
+})
